pipeline/requirements.txt
# Pipeline dependencies
# Core
pydantic>=2.0
python-dotenv

# Stage 1: Scraping
requests
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Stage 2: PDF extraction (install in Colab, NOT locally)
# docling[vlm]

# Stage 3-5: API calls
httpx
tenacity
tiktoken
